requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
streamlit
requests
